+++ src/main/java/uk/gov/hmcts/reform/divorce/caseformatterservice/domain/model/ccd/DnRefusalCaseData.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
+import uk.gov.hmcts.reform.divorce.model.ccd.Document;
 
 import java.util.List;
