--- src/main/java/uk/gov/hmcts/reform/divorce/caseformatterservice/domain/model/ccd/Document.java
@@ -1,26 +1,0 @@
-package uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.ccd;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
-@Data
-public class Document {
-
-    @JsonProperty("DocumentType")
-    private String documentType;
-
-    @JsonProperty("DocumentLink")
-    private DocumentLink documentLink;
-
-    @JsonProperty("DocumentDateAdded")
-    private String documentDateAdded;
-
-    @JsonProperty("DocumentComment")
-    private String documentComment;
-
-    @JsonProperty("DocumentFileName")
-    private String documentFileName;
-
-    @JsonProperty("DocumentEmailContent")
-    private String documentEmailContent;
-}
