+++ src/test/java/uk/gov/hmcts/reform/divorce/caseformatterservice/domain/model/ccd/DocumentUTest.java
@@ -6,6 +6,8 @@ import com.fasterxml.jackson.databind.ObjectWriter;
 import org.junit.Before;
 import org.junit.Test;
 import uk.gov.hmcts.reform.divorce.caseformatterservice.mapper.ObjectMapperTestUtil;
+import uk.gov.hmcts.reform.divorce.model.ccd.Document;
+import uk.gov.hmcts.reform.divorce.model.ccd.DocumentLink;
 
 import java.text.SimpleDateFormat;
 import java.util.Locale;
