+++ src/test/java/uk/gov/hmcts/reform/divorce/caseformatterservice/mapper/DocumentCollectionDocumentRequestMapperUTest.java
@@ -8,8 +8,8 @@ import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.util.ReflectionTestUtils;
 import uk.gov.hmcts.reform.divorce.caseformatterservice.CaseFormatterServiceApplication;
 import uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.ccd.CollectionMember;
-import uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.ccd.Document;
 import uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.documentupdate.GeneratedDocumentInfo;
+import uk.gov.hmcts.reform.divorce.model.ccd.Document;
 
 import static org.junit.Assert.assertEquals;
